Add the possiblity to feed additional context to the logger
It is now possible, through the use of the LLA_LOG_CTX("...")
macro to push some context to the logger.
This helps figuring out the origin and the reason for log
messages.

This is configurable via an option (that defaults to false)
in the configuration file: config.log.context

It is also possible to colorize the log output (useful
when reading them from a Linux terminal). The config
option is config.log.colorize.