Implement support for Islog Key Server.
Islog Key Server (IKS) is a tool that provide remote cryptographic
operation, in order to not store cryptographic key locally.

This (big) commit implements client-side support for IKS.
To delegate cryptographic operation to the IKS, a new KeyStorage
subtype has been introduced: IKSStorage. It stores the name
of the key (instead of its content).

Cryptographic operation for DESFireEv1 will use IKS code path
if it detects that the key used have a IKSStorage attached.

iks::IslogKeyServer is the main class that provide network
support to talk to the IKS.