Minor cleanup of CreateChip
ReaderUnit gains a new protected method to create a default
result checker: createDefaultResultChecker().
It lets each reader unit provide a default implementation
of a ResultChecker object that is suitable for most use-case.

The createChip method in PCSCReaderUnit will still
search for combination of reader/card that may need a custom
ResultChecker. In that case, the ResultChecker provided
by createDefaultResultChecker() will not be used.