Infrastructure change to allow reader identification via USB info.
The LibraryManager::getReader and
PCSCReaderUnit::createPCSCReaderUnit gains an additional parameter.
This parameter is known as the `readerIdentifier`. It can either
be equal to the `readerName` (behavior will not change) or be
set to "usb_identity::${VENDOR_ID}_${PRODUCT_ID}".

This parameter is passed to the C function "getReaderUnit" exposed
by some plugins.

The usb identity string is built using the new static function
PCSCReaderUnit::getUSBInfo().