Add a ReaderFactoryHelper that is used by PCSC plugins.
This new class let us register factory function based
on regular expression matching the reader name or its USB
vendor_id/product_id.

The current code will try to use the USB information to create
the reader unit (if available) and fallback to retrying
with the reader name if it failed.