Update IKS configuration logic.
IKS configuration can now be attached to IKS object to ease
serialization/deserialization of the configuration.

Upon deserialization, if a valid configuration if found, it is
used to call IslogKeyServer::configureGlobalInstance().

On serialization, is the IKSStorage is marked as "want to save
IKS config", the configuration is serialized along the object.