Add support for EPassport (1/2).
This commit and the following one implement support for EPassport.
The implementation is provided by a new card plugin: epass.
    + Basic Access Control is supported, aswel as reading the DataGroup
1 and 2: this gives access to card holder's identity information
aswel as a picture.
    + Validation of data's integrity is not yet implemented.

A new service, IdentityCardService has been introduced.
    + This service is designed to work with various documents that
      would hold information about the card holder's indentity:
      driving license, passport, national identity card.

A new access info object, EPassAccessInfo, has been implemented.
It is used by the Identity service to provide
the service implementation with the MRZ data,
in order to perform authentication.

Note:
    + GTest-backed unit tests are provided regarding the
      cryptographic code.
    + A test for the new Identity card service is also available.