Improve support for CL1356+. WaitInsertion process change.
Wanted behavior: If a card type is forced, waitInsertion() for
ID3ReaderUnit will attempt to find a card whose type match the
forced type.

This require a somewhat custom implement, however waitInsertion()
cannot be overriden, because it is the step at which we create
the proxy reader unit.

A new hook, process_insertion() has been introduced. It is called
by PCSCReaderUnit::waitInsertion() upon dection of a card.
The ID3ReaderUnit override this hook to look for a card that match
the forced type (if any).