Fix loadKey issue with Mifare Classic + ACS122U
This is a potential fix for #107.
The issue lies with using non-volating memory to load the key.
This is not possible on this reader, and an error code is returned.

A similar issue was encountered with ACS1222L and Sony RC-S380.
In order for the MifarePCSCommand code to run properly in that case
this commit changes the LLA error-code associated with PCSC code
0x6300 on this reader to UNKOWN_ERROR.