Add GetPACSBit support for SEOS cards (5x27)
HID GetPACSBit was already supported for iClass / Prox card and
this add support for SEOS card aswell.

SEOS chip will attempt to delegate to the LibraryManager to
retrieve an AccessControl service. If the reader is 5x27 and
proper library is available, a service able to issue hardware
GetPACSBit command will be returned.

Maintance note: Renamed ProxCommand to DummyCommand because
they are now use by both Prox and SEOS chips.