Introduce a new JSON_DUMP service.
The goal of this new service is to provide a very high-level way
of interacting with a card through the LLA library.

The service accept a JSON string as input which provides
information about what to read, and where to read from.
The output of the service is a json object that contains the dump
of the card according to the input specifications.