Fix Desfire Storage Service when using read/write key 0
When using a write key number 0, the key would not be loaded into the DESFireCrypto object, resulting in the default DES key being used instead of the specified AES key.

This commit always loads the key into the DESFireCrypto object, even if the key number is 0.
In addition, there is now a sanity check to ensure that if a read/write key with number 0 and a MasterAppKey are specified, they are equals.