DESFire: Fix combination of key diversification and random UID.
Key diversification relied on card UID.
When random UID is enabled we retrieve the real card UID after
a successful authentication.

However, the initDiversification() call uses the identifier stored
in the desfire crypto object: the identifier is not updated afte
a successful authentication and therefore diversification is done
with a random UID.

This commit simply updates the crypto object identifier after
a successful authentication.